(apollo) load unique schema
The loadFileSync function of @graphql-tools allow us to read multiples files.
But thoses files could not be determined at build time, so they are not included in the output.
This result with an empty schema on lambda execution.
Importing directly the schema allow webpack to include the file in the produced bundle.

This is not the ideal workaround, but it work